Add handler + endpoint to updateProperty
This change allows the main/entrypoint handler to asynchronously invoke
an instance of the updateProperty function for every property. The
benefit here is that the updates can all run and die independently.

This allows each instance to use its own resources because each update
instance is its own stateless function.